sonar sensors: set fov to 40° as stated in http://picaxe.hobbizine.com/srf05.html
"A proper understanding of the detection zone of the SRF05 is key to
developing a successful detection and avoidance algorithm. The beam
width (or beam angle) of ultrasonic range finders is typically described
as being a cone of a certain angle (for the SRF05 it is about 55°). This
angle describes the arc at which the ultrasound pulse emanates from the
transducer. The mistake is sometimes made to consider the detection zone
of the sensor as being identical to this beam. The conclusion of this
logic would be that the detection zone extends from the sensor in an
even, steadily expanding arc until it reaches the 4 meter limit of the
SRF05 range. In reality the detection zone expands at about 55° for the
first meter and then the rate of expansion starts to decay. At about 2
meters distance from the sensor the angle of the detection zone is
closer to 40° with the zone reaching a maximum width of about 80-100
centimeters. From this point the detection zone will begin to narrow,
ultimately reaching zero when the distance from the sensor is a little
more than 4 meters."